OPT: corrected code fragments after merging
Former-commit-id: 562af9a192de32088847ad72d159e5fb240f7d23
Former-commit-id: 9c958c349b21ca3c0492d4428f9b0b5256f31f2a

+++ scripts/sct_maths.py
@@ -63,7 +63,7 @@ def get_parser():
                       description="Threshold image using percentile of its histogram.",
                       mandatory=False)
     parser.add_option(name="-thr",
-                      type_value='int',
+                      type_value='float',
                       description='Use following number to threshold image (zero below number).',
                       mandatory=False,
                       example="")
